refactor code to UserInfo component and use in UserPage
